* sum return aggregate changed to float

+++ VCalculateApi/Interfaces/IStorageClient.cs
@@ -4,6 +4,6 @@ namespace VCalculateApi.Interfaces
 {
     public interface IStorageClient
     {
-        Task<(int sum, float avg)> RetrieveData(string name, int? since, int? to);
+        Task<(float sum, float avg)> RetrieveData(string name, int? since, int? to);
     }
 }
